refactor(session): shared attachment-ingest module + bytes ingest helper
